Add sorting and filtering in Watchlist module
* Make StockTable, StockTableFilters reusable
* Use StockTable, StockTableFilters in Watchlist module
* Refetch favorites after adding a stock to favorites
* Fix React Router path for API requests
