+++ packages/frontend/src/components/etc/StockTable/index.ts
@@ -1,0 +1,1 @@
+export { StockTable } from "./StockTable";
