+++ packages/frontend/src/components/etc/StockTableFilters/index.ts
@@ -1,0 +1,2 @@
+export { StockTableFilters } from "./StockTableFilters";
+export type { StockTableFiltersProps } from "./StockTableFilters";
